Refactoring to codeception callable step
