+++ tests/cli/CliGuy.php
@@ -10,6 +10,7 @@ use Codeception\Module\CliHelper;
 
 /**
  * Inherited methods
+ * @method void run($callable)
  * @method void wantToTest($text)
  * @method void wantTo($text)
  * @method void amTesting($method)
