+++ tests/unit/CodeGuy.php
@@ -10,6 +10,7 @@ use Codeception\Module\EmulateModuleHelper;
 
 /**
  * Inherited methods
+ * @method void run($callable)
  * @method void wantToTest($text)
  * @method void wantTo($text)
  * @method void amTesting($method)
